JNG-5380 Rename Bootstrap To Guice
