+++ judo-runtime-core-jsl-guice/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>judo-runtime-core-bootstrap-jsl</name>
+	<name>judo-runtime-core-guice-jsl</name>
 	<comment></comment>
 	<projects>
 	</projects>
