+++ judo-runtime-core-jsl-guice/pom.xml
@@ -7,10 +7,10 @@
         <version>${revision}</version>
     </parent>
 
-    <artifactId>judo-runtime-core-bootstrap-jsl</artifactId>
+    <artifactId>judo-runtime-core-jsl-guice</artifactId>
     <packaging>bundle</packaging>
 
-    <name>JUDO Runtime Core :: JUDO Language Specification DSL bootstrap</name>
+    <name>JUDO Runtime Core :: JUDO Language Specification DSL guice</name>
 
     <build>
         <plugins>
@@ -19,7 +19,7 @@
                 <artifactId>maven-bundle-plugin</artifactId>
                 <configuration>
                     <instructions>
-                        <Export-Package>hu.blackbelt.judo.runtime.core.bootstrap.jsl.*;version=${project.version}</Export-Package>
+                        <Export-Package>hu.blackbelt.judo.runtime.core.guice.jsl.*;version=${project.version}</Export-Package>
                     </instructions>
                 </configuration>
             </plugin>
@@ -79,7 +79,7 @@
 
         <dependency>
             <groupId>hu.blackbelt.judo.runtime</groupId>
-            <artifactId>judo-runtime-core-bootstrap</artifactId>
+            <artifactId>judo-runtime-core-guice</artifactId>
         </dependency>
 
         <dependency>
