+++ judo-runtime-core-jsl-guice/src/main/java/hu/blackbelt/judo/runtime/core/bootstrap/jsl/JSLTransformationModelLoader.java
@@ -1,4 +1,4 @@
-package hu.blackbelt.judo.runtime.core.bootstrap.jsl;
+package hu.blackbelt.judo.runtime.core.guice.jsl;
 
 /*-
  * #%L
