+++ judo-runtime-core-jsl-guice/src/test/resources/models/sample.jsl
