+++ judo-runtime-core-jsl-itest/common/src/main/java/hu/blackbelt/judo/runtime/core/jsl/modules/JudoPostgresqlTestModules.java
@@ -24,9 +24,9 @@ package hu.blackbelt.judo.runtime.core.jsl.modules;
 import com.google.inject.Singleton;
 import com.google.inject.name.Names;
 import hu.blackbelt.judo.dispatcher.api.Sequence;
-import hu.blackbelt.judo.runtime.core.bootstrap.JudoModule;
-import hu.blackbelt.judo.runtime.core.bootstrap.dao.rdbms.PlatformTransactionManagerProvider;
-import hu.blackbelt.judo.runtime.core.bootstrap.dao.rdbms.postgresql.*;
+import hu.blackbelt.judo.runtime.core.guice.JudoModule;
+import hu.blackbelt.judo.runtime.core.guice.dao.rdbms.PlatformTransactionManagerProvider;
+import hu.blackbelt.judo.runtime.core.guice.dao.rdbms.postgresql.*;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.Dialect;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.RdbmsInit;
 import hu.blackbelt.judo.runtime.core.dao.rdbms.RdbmsParameterMapper;
