+++ judo-runtime-core-jsl-itest/models/pom.xml
@@ -140,12 +140,12 @@
 
         <dependency>
             <groupId>hu.blackbelt.judo.runtime</groupId>
-            <artifactId>judo-runtime-core-bootstrap-hsqldb</artifactId>
+            <artifactId>judo-runtime-core-guice-hsqldb</artifactId>
         </dependency>
 
         <dependency>
             <groupId>hu.blackbelt.judo.runtime</groupId>
-            <artifactId>judo-runtime-core-bootstrap-postgresql</artifactId>
+            <artifactId>judo-runtime-core-guice-postgresql</artifactId>
         </dependency>
 
         <dependency>
