+++ pom.xml
@@ -41,7 +41,7 @@
         </osgi-transaction-import>
 
         <judo-tatami-jsl-version>1.1.4.20231220_192655_997b9b34_develop</judo-tatami-jsl-version>
-        <judo-runtime-core-version>1.0.6.20231213_134008_c0a2ca94_develop</judo-runtime-core-version>
+        <judo-runtime-core-version>1.0.6.20231229_231036_76df17fd_feature_JNG_5380_RenameBootstrapToGuice</judo-runtime-core-version>
         <judo-psm-generator-sdk-core-version>1.0.0.20231130_090851_402cfa2a_develop</judo-psm-generator-sdk-core-version>
 
         <judo-meta-psm-version>1.3.0.20231107_041329_29cfda41_develop</judo-meta-psm-version>
@@ -584,7 +584,7 @@
             </activation>
             <modules>
                 <module>judo-runtime-core-jsl</module>
-                <module>judo-runtime-core-bootstrap-jsl</module>
+                <module>judo-runtime-core-jsl-guice</module>
                 <module>judo-runtime-core-jsl-itest</module>
                 <module>judo-runtime-core-jsl-spring-itest</module>
             </modules>
